[feat/daengle-91] 댕글미터 도메인 설계, 댕글미터 비즈니스 로직 구현

+++ daengle-persistence-mysql/src/main/java/ddog/persistence/mysql/jpa/entity/GroomerDanegleMeterJpaEntity.java
@@ -21,7 +21,7 @@ public class GroomerDanegleMeterJpaEntity {
     Long groomerDaengleMeterId;
 
     Long groomerId;
-    Long score;
+    Integer score;
     Long evaluationCount;
 
     public GroomerDaengleMeter toModel() {
